refactor temperature/humidity methods
